+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#123] If the Helm client can't delete resources during a rollback, the installation gets stuck 
+  in the `uninstalling` state. To Prevent this, we reinstall the component.
 
 ## [v1.15.0] - 2026-09-10
 ### Changed